refactored move
now a  "naked" nested class in gamestate
