+++ core/src/au/id/hxb/cathedroid/Mechanics/Piece.java
@@ -11,6 +11,26 @@ public enum Piece {
     L_ST2,
     L_IN1,
     L_IN2,
-
-
+    L_BR,
+    L_SQ,
+    L_AB,
+    L_MA,
+    L_TO,
+    L_IF,
+    L_CS,
+    L_AC,
+    D_TA1,
+    D_TA2,
+    D_ST1,
+    D_ST2,
+    D_IN1,
+    D_IN2,
+    D_BR,
+    D_SQ,
+    D_AB,
+    D_MA,
+    D_TO,
+    D_IF,
+    D_CS,
+    D_AC,
 }
